Reorganize EMS directory
