+++ src/components/EMS/HistogramChart/hooks/useChart.tsx
