+++ src/components/EMS/HistogramChart/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./HistogramChart";
