+++ src/components/EMS/index.tsx
@@ -1,1 +1,2 @@
-export { default } from "./EMS";
+export { default as EMS } from "./EMS";
+export { default as CompletedRun } from "./CompletedRun";
